+++ src/app/app.routes.ts
@@ -76,6 +76,7 @@ import { DataGridRowReorderDemoComponent } from './datagrid/datagrid-rowreorder.
 import { DataGridDirtyIndicationDemoComponent } from './datagrid/datagrid-dirty-indication.demo';
 import { DataGridDynamicDemoComponent } from './datagrid/datagrid-dynamic.demo';
 import { DataGridAddRowDemoComponent } from './datagrid/datagrid-add-row.demo';
+import { DataGridLookupCellChangeDemoComponent } from './datagrid/datagrid-lookup-cell-change.demo';
 import { DataGridEditorsDemoComponent } from './datagrid/datagrid-editors.demo';
 import { DataGridEmptyMessageDemoComponent } from './datagrid/datagrid-empty-message.demo';
 import { DataGridExportWithoutDataGridDemoComponent } from './datagrid/datagrid-export-without-datagrid.demo';
@@ -326,6 +327,7 @@ export const routes: Routes = [
   { path: 'datagrid-dirty-indication', component: DataGridDirtyIndicationDemoComponent },
   { path: 'datagrid-dynamic', component: DataGridDynamicDemoComponent },
   { path: 'datagrid-add-row', component: DataGridAddRowDemoComponent },
+  { path: 'datagrid-lookup-cell-change', component: DataGridLookupCellChangeDemoComponent },
   { path: 'datagrid-empty-message', component: DataGridEmptyMessageDemoComponent },
   { path: 'datagrid-editors', component: DataGridEditorsDemoComponent },
   { path: 'datagrid-editors', component: DataGridEditorsDemoComponent },